+++ packages/react-native/ReactCommon/react/renderer/graphics/ColorStop.cpp
@@ -12,7 +12,7 @@ namespace facebook::react {
 #ifdef RN_SERIALIZABLE_STATE
 folly::dynamic ColorStop::toDynamic() const {
   folly::dynamic result = folly::dynamic::object();
-  result["color"] = *color;
+  result["color"] = static_cast<int32_t>(*color);
   result["position"] = position.toDynamic();
   return result;
 }
